+++ src/main/java/com/black_dog20/servertabinfo/client/ClientDataManager.java
@@ -1,14 +1,10 @@
 package com.black_dog20.servertabinfo.client;
 
 import com.black_dog20.servertabinfo.common.utils.Dimension;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.TextComponent;
+import net.minecraft.network.chat.BaseComponent;
+import net.minecraft.resources.ResourceLocation;
 
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
 public class ClientDataManager {
     public static boolean modOnServer = false;
@@ -16,5 +12,5 @@ public class ClientDataManager {
     public static List<Dimension> DIMENSIONS = new LinkedList<>();
     public static Map<UUID, ResourceLocation> PLAYER_DIMENSIONS = new HashMap<>();
 
-    public static Map<ResourceLocation, TextComponent> DIMENSION_NAME_CACHE = new HashMap<>();
+    public static Map<ResourceLocation, BaseComponent> DIMENSION_NAME_CACHE = new HashMap<>();
 }