+++ src/main/java/com/black_dog20/servertabinfo/client/ClientStartup.java
@@ -7,9 +7,9 @@ import com.black_dog20.servertabinfo.client.overlays.PlayerListOverlay;
 import com.black_dog20.servertabinfo.client.overlays.TpsListOverlay;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.client.registry.ClientRegistry;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fmlclient.registry.ClientRegistry;
 
 @Mod.EventBusSubscriber( modid = ServerTabInfo.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public class ClientStartup {