+++ src/main/java/com/black_dog20/servertabinfo/client/overlays/PlayerListOverlay.java
@@ -1,7 +1,7 @@
 package com.black_dog20.servertabinfo.client.overlays;
 
 import com.black_dog20.bml.client.DrawingContext;
-import com.black_dog20.bml.client.overlay.Overlay;
+import com.black_dog20.bml.client.overlay.GameOverlay;
 import com.black_dog20.bml.client.rows.Row;
 import com.black_dog20.bml.client.rows.RowHelper;
 import com.black_dog20.bml.client.rows.columns.BlankColumn;
@@ -17,24 +17,24 @@ import com.black_dog20.servertabinfo.client.keybinds.Keybinds;
 import com.black_dog20.servertabinfo.common.utils.Dimension;
 import com.google.common.collect.ComparisonChain;
 import com.google.common.collect.Ordering;
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.ChatFormatting;
+import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.network.play.ClientPlayNetHandler;
-import net.minecraft.client.network.play.NetworkPlayerInfo;
-import net.minecraft.client.renderer.ItemRenderer;
-import net.minecraft.client.world.ClientWorld;
-import net.minecraft.scoreboard.Score;
-import net.minecraft.scoreboard.ScoreObjective;
-import net.minecraft.scoreboard.ScorePlayerTeam;
-import net.minecraft.scoreboard.Scoreboard;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.Util;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.TextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.world.GameType;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.multiplayer.ClientPacketListener;
+import net.minecraft.client.multiplayer.PlayerInfo;
+import net.minecraft.client.renderer.entity.ItemRenderer;
+import net.minecraft.network.chat.BaseComponent;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.GameType;
+import net.minecraft.world.scores.Objective;
+import net.minecraft.world.scores.PlayerTeam;
+import net.minecraft.world.scores.Score;
+import net.minecraft.world.scores.Scoreboard;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.event.RenderGameOverlayEvent;
@@ -49,33 +49,33 @@ import java.util.stream.Collectors;
 import static com.black_dog20.servertabinfo.common.utils.Translations.*;
 
 @OnlyIn(Dist.CLIENT)
-public class PlayerListOverlay extends Overlay.Pre {
+public class PlayerListOverlay extends GameOverlay.Pre {
 
-    private static final Ordering<NetworkPlayerInfo> ENTRY_ORDERING = Ordering.from(new PlayerComparator());
+    private static final Ordering<PlayerInfo> ENTRY_ORDERING = Ordering.from(new PlayerComparator());
     private Minecraft minecraft;
-    private FontRenderer fontRenderer;
+    private Font fontRenderer;
     private ItemRenderer itemRenderer;
-    private long lastRenderTime = Util.milliTime();
+    private long lastRenderTime = Util.getMillis();
     private int ticks = 0;
     private int page = 1;
 
     public PlayerListOverlay() {
         this.minecraft = Minecraft.getInstance();
-        this.fontRenderer = minecraft.fontRenderer;
+        this.fontRenderer = minecraft.font;
         this.itemRenderer = minecraft.getItemRenderer();
     }
 
     @Override
-    public void onRender(MatrixStack matrixStack, int width, int height) {
-        if(Keybinds.SHOW.isKeyDown())
+    public void onRender(PoseStack matrixStack, int width, int height) {
+        if(Keybinds.SHOW.isDown())
             return;
         int y = 10;
         int z = 0;
-        if (Util.milliTime() - 2000 > lastRenderTime) {
+        if (Util.getMillis() - 2000 > lastRenderTime) {
             page = 1;
             ticks = 1;
         }
-        int itemsPerPage = (int) Math.floor((height - 7 * y) / fontRenderer.FONT_HEIGHT);
+        int itemsPerPage = (int) Math.floor((height - 7 * y) / fontRenderer.lineHeight);
         List<Row> rows = getRows();
 
         int maxPages = (int) Math.ceil(rows.size() / (double) itemsPerPage);
@@ -94,9 +94,9 @@ public class PlayerListOverlay extends Overlay.Pre {
 
         DrawingContext drawingContext = new DrawingContext(matrixStack, width, height, x, y, z, fontRenderer, itemRenderer);
         y = RowHelper.drawRowsWithBackground(drawingContext, rows);
-        fontRenderer.func_243246_a(matrixStack, PAGE.get(page, maxPages), width / 2 + 2, y + 2, -1);
+        fontRenderer.drawShadow(matrixStack, PAGE.get(page, maxPages), width / 2 + 2, y + 2, -1);
         ticks++;
-        lastRenderTime = Util.milliTime();
+        lastRenderTime = Util.getMillis();
     }
 
     @Override
@@ -114,18 +114,18 @@ public class PlayerListOverlay extends Overlay.Pre {
     }
 
     private List<Row> getRows() {
-        ClientPlayNetHandler nethandlerplayclient = this.minecraft.player.connection;
-        List<NetworkPlayerInfo> list = ENTRY_ORDERING.<NetworkPlayerInfo>sortedCopy(nethandlerplayclient.getPlayerInfoMap());
+        ClientPacketListener nethandlerplayclient = this.minecraft.player.connection;
+        List<PlayerInfo> list = ENTRY_ORDERING.<PlayerInfo>sortedCopy(nethandlerplayclient.getOnlinePlayers());
 
         return list.stream()
                 .map(this::buildRow)
                 .collect(Collectors.toCollection(LinkedList::new));
     }
 
-    private Row buildRow(NetworkPlayerInfo playerInfo) {
+    private Row buildRow(PlayerInfo playerInfo) {
         Row.RowBuilder builder = new Row.RowBuilder();
-        Scoreboard scoreboard = Optional.ofNullable(minecraft.world)
-                .map(ClientWorld::getScoreboard)
+        Scoreboard scoreboard = Optional.ofNullable(minecraft.level)
+                .map(ClientLevel::getScoreboard)
                 .orElse(null);
 
         return builder
@@ -141,37 +141,37 @@ public class PlayerListOverlay extends Overlay.Pre {
                 .build();
     }
 
-    private ITextComponent getPlayerScore(NetworkPlayerInfo playerInfo, Scoreboard scoreboard) {
-        ScoreObjective scoreObjective = Optional.ofNullable(scoreboard)
-                .map(s -> s.getObjectiveInDisplaySlot(0))
+    private BaseComponent getPlayerScore(PlayerInfo playerInfo, Scoreboard scoreboard) {
+        Objective scoreObjective = Optional.ofNullable(scoreboard)
+                .map(s -> s.getDisplayObjective(0))
                 .orElse(null);
 
         return TextComponentBuilder.of("")
                 .with(getScore(playerInfo, scoreboard, scoreObjective), showPlayerScore(playerInfo, scoreObjective))
-                .format(TextFormatting.YELLOW)
+                .format(ChatFormatting.YELLOW)
                 .build();
     }
 
-    private Supplier<Boolean> showPlayerScore(NetworkPlayerInfo playerInfo, ScoreObjective scoreObjective) {
-        return () -> scoreObjective != null && GameType.SPECTATOR != playerInfo.getGameType();
+    private Supplier<Boolean> showPlayerScore(PlayerInfo playerInfo, Objective scoreObjective) {
+        return () -> scoreObjective != null && GameType.SPECTATOR != playerInfo.getGameMode();
     }
 
-    private int getScore(NetworkPlayerInfo playerInfo, Scoreboard scoreboard, ScoreObjective scoreObjective) {
+    private int getScore(PlayerInfo playerInfo, Scoreboard scoreboard, Objective scoreObjective) {
         if (scoreObjective != null) {
-            Score score = scoreboard.getOrCreateScore(playerInfo.getGameProfile().getName(), scoreObjective);
-            return score.getScorePoints();
+            Score score = scoreboard.getOrCreatePlayerScore(playerInfo.getProfile().getName(), scoreObjective);
+            return score.getScore();
         } else {
             return 0;
         }
     }
 
     private Supplier<Boolean> showScoreObjective(Scoreboard scoreboard) {
-        return () -> scoreboard != null && scoreboard.getObjectiveInDisplaySlot(0) != null;
+        return () -> scoreboard != null && scoreboard.getDisplayObjective(0) != null;
     }
 
     private boolean isAllowed() {
-        return !Config.OP_ONLY_MODE.get() || minecraft.isSingleplayer() || Optional.ofNullable(minecraft.player)
-                .map(player -> player.hasPermissionLevel(1))
+        return !Config.OP_ONLY_MODE.get() || minecraft.hasSingleplayerServer() || Optional.ofNullable(minecraft.player)
+                .map(player -> player.hasPermissions(1))
                 .orElse(false);
     }
 
@@ -183,55 +183,55 @@ public class PlayerListOverlay extends Overlay.Pre {
         return rows;
     }
 
-    private ITextComponent getPlayerName(NetworkPlayerInfo playerInfo) {
+    private Component getPlayerName(PlayerInfo playerInfo) {
         return Optional.ofNullable(playerInfo)
-                .map(NetworkPlayerInfo::getDisplayName)
-                .orElseGet(() -> ScorePlayerTeam.func_237500_a_(playerInfo.getPlayerTeam(), new StringTextComponent(playerInfo.getGameProfile().getName())));
+                .map(PlayerInfo::getTabListDisplayName)
+                .orElseGet(() -> PlayerTeam.formatNameForTeam(playerInfo.getTeam(), new TextComponent(playerInfo.getProfile().getName())));
     }
 
-    private ITextComponent getPlayerPing(NetworkPlayerInfo playerInfo) {
-        return TextComponentBuilder.of(playerInfo.getResponseTime())
+    private BaseComponent getPlayerPing(PlayerInfo playerInfo) {
+        return TextComponentBuilder.of(playerInfo.getLatency())
                 .with(MS)
                 .build();
     }
 
-    private ITextComponent getPlayerDim(NetworkPlayerInfo playerInfo) {
-        ResourceLocation dimName = ClientDataManager.PLAYER_DIMENSIONS.getOrDefault(playerInfo.getGameProfile().getId(), null);
+    private BaseComponent getPlayerDim(PlayerInfo playerInfo) {
+        ResourceLocation dimName = ClientDataManager.PLAYER_DIMENSIONS.getOrDefault(playerInfo.getProfile().getId(), null);
         return TextComponentBuilder.of(getDimensionName(dimName))
                 .with(getDimensionTps(dimName), this::isAllowed)
                 .build();
 
     }
 
-    private TextComponent getDimensionName(ResourceLocation dimensionName) {
+    private BaseComponent getDimensionName(ResourceLocation dimensionName) {
         if (dimensionName == null)
-            return (TextComponent) UNKOWN.get();
+            return (BaseComponent) UNKOWN.get();
 
         if (ClientDataManager.DIMENSION_NAME_CACHE.containsKey(dimensionName))
             return ClientDataManager.DIMENSION_NAME_CACHE.get(dimensionName);
 
-        TextComponent name = DimensionUtil.getFormattedDimensionName(dimensionName, ServerTabInfo.MOD_ID);
+        BaseComponent name = DimensionUtil.getFormattedDimensionName(dimensionName, ServerTabInfo.MOD_ID);
         ClientDataManager.DIMENSION_NAME_CACHE.put(dimensionName, name);
         return name;
     }
 
-    private TextComponent getDimensionTps(ResourceLocation dimensionName) {
+    private BaseComponent getDimensionTps(ResourceLocation dimensionName) {
         if (dimensionName == null)
-            return new StringTextComponent("");
+            return new TextComponent("");
         Dimension dimension = ClientDataManager.DIMENSIONS.stream()
                 .filter(d -> d.name.equals(dimensionName))
                 .findFirst()
                 .orElse(null);
         if (dimension == null)
-            return new StringTextComponent("");
+            return new TextComponent("");
 
         int tps = dimension.tps;
-        TextFormatting color = TextFormatting.GREEN;
+        ChatFormatting color = ChatFormatting.GREEN;
 
         if (tps < 20)
-            color = TextFormatting.YELLOW;
+            color = ChatFormatting.YELLOW;
         if (tps <= 10)
-            color = TextFormatting.RED;
+            color = ChatFormatting.RED;
 
         return TextComponentBuilder.of("(")
                 .with(tps)
@@ -241,21 +241,21 @@ public class PlayerListOverlay extends Overlay.Pre {
     }
 
     @OnlyIn(Dist.CLIENT)
-    static class PlayerComparator implements Comparator<NetworkPlayerInfo> {
+    static class PlayerComparator implements Comparator<PlayerInfo> {
 
-        public int compare(NetworkPlayerInfo player1, NetworkPlayerInfo player2) {
-            ScorePlayerTeam scoreplayerteam1 = player1.getPlayerTeam();
-            ScorePlayerTeam scoreplayerteam2 = player2.getPlayerTeam();
+        public int compare(PlayerInfo player1, PlayerInfo player2) {
+            PlayerTeam scoreplayerteam1 = player1.getTeam();
+            PlayerTeam scoreplayerteam2 = player2.getTeam();
             return ComparisonChain.start()
-                    .compareTrueFirst(player1.getGameType() != GameType.SPECTATOR, player2.getGameType() != GameType.SPECTATOR)
+                    .compareTrueFirst(player1.getGameMode() != GameType.SPECTATOR, player2.getGameMode() != GameType.SPECTATOR)
                     .compare(getName(scoreplayerteam1), getName(scoreplayerteam2))
-                    .compare(player1.getGameProfile().getName(), player2.getGameProfile().getName())
+                    .compare(player1.getProfile().getName(), player2.getProfile().getName())
                     .result();
         }
 
-        private String getName(ScorePlayerTeam scorePlayerTeam) {
+        private String getName(PlayerTeam scorePlayerTeam) {
             return Optional.ofNullable(scorePlayerTeam)
-                    .map(ScorePlayerTeam::getName)
+                    .map(PlayerTeam::getName)
                     .orElse("");
         }
     }