+++ src/main/java/com/black_dog20/servertabinfo/client/overlays/TpsListOverlay.java
@@ -1,7 +1,7 @@
 package com.black_dog20.servertabinfo.client.overlays;
 
 import com.black_dog20.bml.client.DrawingContext;
-import com.black_dog20.bml.client.overlay.Overlay;
+import com.black_dog20.bml.client.overlay.GameOverlay;
 import com.black_dog20.bml.client.rows.Row;
 import com.black_dog20.bml.client.rows.RowHelper;
 import com.black_dog20.bml.client.rows.columns.BlankColumn;
@@ -15,17 +15,18 @@ import com.black_dog20.servertabinfo.client.ClientDataManager;
 import com.black_dog20.servertabinfo.client.keybinds.Keybinds;
 import com.black_dog20.servertabinfo.common.utils.Dimension;
 import com.google.common.collect.ImmutableList;
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.ChatFormatting;
+import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.renderer.ItemRenderer;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.Util;
-import net.minecraft.util.text.TextComponent;
-import net.minecraft.util.text.TextFormatting;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.renderer.entity.ItemRenderer;
+import net.minecraft.network.chat.BaseComponent;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.client.event.RenderGameOverlayEvent;
+import net.minecraftforge.client.gui.ForgeIngameGui;
+import net.minecraftforge.client.gui.IIngameOverlay;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -35,24 +36,24 @@ import java.util.stream.Collectors;
 import static com.black_dog20.servertabinfo.common.utils.Translations.*;
 
 @OnlyIn(Dist.CLIENT)
-public class TpsListOverlay extends Overlay.Pre {
+public class TpsListOverlay extends GameOverlay.PreLayer {
 
     private Minecraft minecraft;
-    private FontRenderer fontRenderer;
+    private Font fontRenderer;
     private ItemRenderer itemRenderer;
-    private long lastRenderTime = Util.milliTime();
+    private long lastRenderTime = Util.getMillis();
     private int ticks = 0;
     private int page = 1;
     private boolean hasBeenShown = false;
 
     public TpsListOverlay() {
         this.minecraft = Minecraft.getInstance();
-        this.fontRenderer = minecraft.fontRenderer;
+        this.fontRenderer = minecraft.font;
         this.itemRenderer = minecraft.getItemRenderer();
     }
 
     @Override
-    public void onRender(MatrixStack matrixStack, int width, int height) {
+    public void onRender(PoseStack matrixStack, int width, int height) {
         int y = 10;
         int z = 0;
 
@@ -63,12 +64,12 @@ public class TpsListOverlay extends Overlay.Pre {
         }
     }
 
-    private void renderInstalled(MatrixStack matrixStack, int width, int height, int y, int z) {
-        if (Util.milliTime() - 2000 > lastRenderTime) {
+    private void renderInstalled(PoseStack matrixStack, int width, int height, int y, int z) {
+        if (Util.getMillis() - 2000 > lastRenderTime) {
             page = 1;
             ticks = 1;
         }
-        int itemsPerPage = (int) Math.floor((height - 7 * y) / fontRenderer.FONT_HEIGHT);
+        int itemsPerPage = (int) Math.floor((height - 7 * y) / fontRenderer.lineHeight);
         List<Row> rows = getRows();
 
         int maxPages = (int) Math.ceil(rows.size() / (double) itemsPerPage);
@@ -87,12 +88,12 @@ public class TpsListOverlay extends Overlay.Pre {
 
         DrawingContext drawingContext = new DrawingContext(matrixStack, width, height, x, y, z, fontRenderer, itemRenderer);
         y = RowHelper.drawRowsWithBackground(drawingContext, rows);
-        fontRenderer.func_243246_a(matrixStack, PAGE.get(page, maxPages), width / 2 + 2, y + 2, -1);
+        fontRenderer.drawShadow(matrixStack, PAGE.get(page, maxPages), width / 2 + 2, y + 2, -1);
         ticks++;
-        lastRenderTime = Util.milliTime();
+        lastRenderTime = Util.getMillis();
     }
 
-    private void renderNotInstalled(MatrixStack matrixStack, int width, int height, int z) {
+    private void renderNotInstalled(PoseStack matrixStack, int width, int height, int z) {
         List<Row> rows = ImmutableList.of(new Row.RowBuilder()
                 .withColumn(ITextComponentColumn.of("text", NOT_INSTALLED.get()))
                 .build());
@@ -105,14 +106,14 @@ public class TpsListOverlay extends Overlay.Pre {
     }
 
     @Override
-    public boolean doRender(RenderGameOverlayEvent.ElementType elementType) {
-        if (elementType == RenderGameOverlayEvent.ElementType.CROSSHAIRS) {
-            if (Keybinds.SHOW.isKeyDown()) {
+    public boolean doRender(IIngameOverlay iIngameOverlay) {
+        if (iIngameOverlay == ForgeIngameGui.CROSSHAIR_ELEMENT) {
+            if (Keybinds.SHOW.isDown()) {
                 boolean allowed = isAllowed();
                 if (!hasBeenShown && !allowed) {
                     hasBeenShown = true;
                     Optional.ofNullable(minecraft.player)
-                            .ifPresent(player -> player.sendStatusMessage(NOT_ALLOWED.get(), true));
+                            .ifPresent(player -> player.displayClientMessage(NOT_ALLOWED.get(), true));
                 }
 
                 return allowed;
@@ -125,8 +126,8 @@ public class TpsListOverlay extends Overlay.Pre {
     }
 
     private boolean isAllowed() {
-        return !Config.OP_ONLY_MODE.get() || minecraft.isSingleplayer() || Optional.ofNullable(minecraft.player)
-                .map(player -> player.hasPermissionLevel(1))
+        return !Config.OP_ONLY_MODE.get() || minecraft.hasSingleplayerServer() || Optional.ofNullable(minecraft.player)
+                .map(player -> player.hasPermissions(1))
                 .orElse(false);
     }
 
@@ -165,18 +166,18 @@ public class TpsListOverlay extends Overlay.Pre {
         return rows;
     }
 
-    private TextComponent getDimensionName(Dimension dimension) {
+    private BaseComponent getDimensionName(Dimension dimension) {
         ResourceLocation dimensionName = dimension.name;
 
         if(ClientDataManager.DIMENSION_NAME_CACHE.containsKey(dimensionName))
             return ClientDataManager.DIMENSION_NAME_CACHE.get(dimensionName);
 
-        TextComponent name = DimensionUtil.getFormattedDimensionName(dimensionName, ServerTabInfo.MOD_ID);
+        BaseComponent name = DimensionUtil.getFormattedDimensionName(dimensionName, ServerTabInfo.MOD_ID);
         ClientDataManager.DIMENSION_NAME_CACHE.put(dimensionName, name);
         return name;
     }
 
-    private TextComponent getDimensionMean(Dimension dimension) {
+    private BaseComponent getDimensionMean(Dimension dimension) {
         double mean = dimension.meanTickTime;
 
         return TextComponentBuilder.of(String.format("%.2f", mean))
@@ -184,14 +185,14 @@ public class TpsListOverlay extends Overlay.Pre {
                 .build();
     }
 
-    private TextComponent getDimensionTps(Dimension dimension) {
+    private BaseComponent getDimensionTps(Dimension dimension) {
         int tps = dimension.tps;
-        TextFormatting color = TextFormatting.GREEN;
+        ChatFormatting color = ChatFormatting.GREEN;
 
         if (tps < 20)
-			color = TextFormatting.YELLOW;
+			color = ChatFormatting.YELLOW;
 		if (tps <= 10)
-            color = TextFormatting.RED;
+            color = ChatFormatting.RED;
 
 		return TextComponentBuilder.of("(")
                 .with(tps)