+++ src/main/java/com/black_dog20/servertabinfo/common/events/ServerEvents.java
@@ -7,21 +7,17 @@ import com.black_dog20.servertabinfo.common.network.packets.PacketDimensions;
 import com.black_dog20.servertabinfo.common.network.packets.PacketPlayers;
 import com.black_dog20.servertabinfo.common.utils.Dimension;
 import com.mojang.datafixers.util.Pair;
-import net.minecraft.entity.player.ServerPlayerEntity;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.world.server.ServerWorld;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.server.ServerLifecycleHooks;
+import net.minecraftforge.fmllegacy.server.ServerLifecycleHooks;
 
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
 @Mod.EventBusSubscriber( modid = ServerTabInfo.MOD_ID)
 public class ServerEvents {
@@ -36,26 +32,26 @@ public class ServerEvents {
             return;
 
         MinecraftServer server = ServerLifecycleHooks.getCurrentServer();
-        if(server.getPlayerList().getCurrentPlayerCount() == 0){
+        if(server.getPlayerList().getPlayerCount() == 0){
             ticks = 0;
             return;
         }
 
         if(ticks % Config.REFRESH_TICKS.get() == 0) {
-            Pair<Integer, Double> overall = getTpsAndMean(server.tickTimeArray);
+            Pair<Integer, Double> overall = getTpsAndMean(server.tickTimes);
             DIMENSIONS.clear();
             DIMENSIONS.add(new Dimension(new ResourceLocation(ServerTabInfo.MOD_ID, "overall"), overall.getSecond(), overall.getFirst()));
-            for (ServerWorld world : server.forgeGetWorldMap().values()) {
-                ResourceLocation name = world.getDimensionKey().getLocation();
-                Pair<Integer, Double> tpsAndMean = getTpsAndMean(server.getTickTime(world.getDimensionKey()));
+            for (ServerLevel world : server.forgeGetWorldMap().values()) {
+                ResourceLocation name = world.dimension().location();
+                Pair<Integer, Double> tpsAndMean = getTpsAndMean(server.getTickTime(world.dimension()));
                 DIMENSIONS.add(new Dimension(name, tpsAndMean.getSecond(), tpsAndMean.getFirst()));
 		    }
 
-            List<ServerPlayerEntity> playerList = server.getPlayerList().getPlayers();
+            List<ServerPlayer> playerList = server.getPlayerList().getPlayers();
             PLAYER_DIMENSIONS.clear();
-            for (ServerPlayerEntity player : playerList) {
-                ResourceLocation name = player.getEntityWorld().getDimensionKey().getLocation();
-                PLAYER_DIMENSIONS.put(player.getUniqueID(), name);
+            for (ServerPlayer player : playerList) {
+                ResourceLocation name = player.getCommandSenderWorld().dimension().location();
+                PLAYER_DIMENSIONS.put(player.getUUID(), name);
             }
 
             PacketHandler.sendToAll(new PacketDimensions(DIMENSIONS));
@@ -84,10 +80,10 @@ public class ServerEvents {
 
     @SubscribeEvent
     public static void onPlayerLogin(PlayerEvent.PlayerLoggedInEvent event) {
-        if (!event.getPlayer().world.isRemote) {
-            ServerPlayerEntity playerEntity = (ServerPlayerEntity) event.getPlayer();
-            ResourceLocation name = playerEntity.getEntityWorld().getDimensionKey().getLocation();
-            PLAYER_DIMENSIONS.put(playerEntity.getUniqueID(), name);
+        if (!event.getPlayer().level.isClientSide) {
+            ServerPlayer playerEntity = (ServerPlayer) event.getPlayer();
+            ResourceLocation name = playerEntity.getCommandSenderWorld().dimension().location();
+            PLAYER_DIMENSIONS.put(playerEntity.getUUID(), name);
 
             PacketHandler.sendTo(new PacketDimensions(DIMENSIONS), playerEntity);
             PacketHandler.sendTo(new PacketPlayers(PLAYER_DIMENSIONS), playerEntity);